Tests: Move test file content to dedicated files
Otherwise this is getting really messy
