+++ tests/files/readme_1_expected.md
@@ -1,0 +1,11 @@
+## Package Status
+
+| Bintray | Appveyor | Travis |
+|---------|------------|--------|
+|[![Download](https://api.bintray.com/packages/bincrafters/public-conan/glog%3Abincrafters/images/download.svg) ](https://bintray.com/bincrafters/public-conan/glog%3Abincrafters/_latestVersion)|[![Build status](https://ci.appveyor.com/api/projects/status/github/bincrafters/conan-glog?svg=true)](https://ci.appveyor.com/project/bincrafters/conan-glog)|[![Build Status](https://travis-ci.com/bincrafters/conan-glog.svg)](https://travis-ci.com/bincrafters/conan-glog)|
+
+## Conan.io Information
+
+Bincrafters packages can be found in the following public Conan repository:
+
+[Bincrafters Public Conan Repository on Bintray](https://bintray.com/bincrafters/public-conan)
